pw-schedule: Use <year-month-day> format (ISO 8601) for dates
Format %m/%d/%Y is ambiguous, given that some parts of the world prefer
to write the day first and the month second. This does not matter much
for the CSV that is to be fed to a caldendar tool, but it might still be
confusing for the help message of the tool, for example.

Use something less ambiguous: ISO 8601 (%Y-%m-%d) is a good candidate.

Also tested that the Google calendar import works fine with this.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>