pw-apply: Emit a warning when applying too fresh patches
I tend to miss the fact that some series may be too fresh to be applied:
let me get more help from the tool.

Add mbox_from_series() the ability to warn if the series is not old
as optionally specified, and set a default threshold of 24 hours for
`pw-apply`.

Add a command line argument (--min-age/-m) to allow setting a different
threshold.

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>