pw-apply: don't count weekend time in how long patch has been on the list
We recently started printing a warning when patches are applied
before they spent 24h on the list. The motivation for that was
to let reviewers from all time zones post their comments.

To further improve reviewers' (chances of) work-life balance
exclude weekend time from the 24h limit, meaning that the
patch has to spent 24h of work day time on the list. This
is useful when patches are posted on the list over the weekend
and even tho they may be 2 days old for someone in the eastern
/ positive timezones the folks in western / negative timezones
most likely haven't seen them.

The "summary" line for a patch which timeline does not "straddle"
weekend time remains unchanged:

  By: Some Person  Age: 1d 11h  Tree: net-next  Version: 1  Patches: 2

but if weekend time was in the span (e.g. 10 hours) we'll see:

  By: Some Person  Age: 1d 1h (+10h)  Tree: net-next  Version: 1  Patches: 2

We could say something like "weekend" next to the plus but its
a long word and my terminals are 80 chars :S

Signed-off-by: Jakub Kicinski <kuba@kernel.org>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>